+++ apps/docs/content/docs/develop/(hooks)/use-theme-toggler.client.tsx
@@ -1,21 +1,19 @@
 "use client";
 
-import { ThemeToggler } from "@govtechmy/myds-react/hooks";
+import { ThemeProvider } from "@govtechmy/myds-react/hooks";
 import { MoonIcon, SunIcon } from "@govtechmy/myds-react/icon";
-
-export {
-  useTheme,
-  ThemeProvider,
-  ThemeToggler,
-} from "@govtechmy/myds-react/hooks";
+import { ThemeToggler } from "@govtechmy/myds-react/theme-toggler";
+export { useTheme, ThemeProvider } from "@govtechmy/myds-react/hooks";
 
 export const SampleThemeToggler = () => {
   return (
-    <ThemeToggler
-      themes={[
-        { theme: "light", icon: <SunIcon /> },
-        { theme: "dark", icon: <MoonIcon /> },
-      ]}
-    />
+    <ThemeProvider>
+      <ThemeToggler
+        themes={[
+          { theme: "light", icon: <SunIcon /> },
+          { theme: "dark", icon: <MoonIcon /> },
+        ]}
+      />
+    </ThemeProvider>
   );
 };
