+++ apps/docs/content/docs/develop/(hooks)/use-theme-toggler.mdx
@@ -8,7 +8,6 @@ import Preview from "@/components/preview";
 import {
   ThemeProvider,
   useTheme,
-  ThemeToggler,
   SampleThemeToggler,
 } from "./use-theme-toggler.client";
 import { TypeTable } from "fumadocs-ui/components/type-table";
@@ -17,10 +16,11 @@ import { Tab, Tabs } from "fumadocs-ui/components/tabs";
 <Tabs items={["Preview", "Code"]} defaultValue="Preview">
   <Tab value="Preview">
     <div className="flex justify-center">
-      <ThemeProvider>
-        <ThemeToggler />
-      </ThemeProvider>
+
+      <SampleThemeToggler/>
+
     </div>
+
   </Tab>
   <Tab value="Code">
     ```tsx
@@ -70,9 +70,6 @@ Wrap your application with `ThemeProvider` and use `ThemeToggler` to switch betw
 <Tabs items={["Preview", "Code"]} defaultValue="Preview">
   <Tab value="Preview">
     <div className="flex justify-center">
-      <ThemeProvider>
-        <ThemeToggler />
-      </ThemeProvider>
     </div>
   </Tab>
   <Tab value="Code">
@@ -93,9 +90,6 @@ Pass an array of themes to `ThemeToggler` to customize the available themes.
 <Tabs items={["Preview", "Code"]} defaultValue="Preview">
   <Tab value="Preview">
     <div className="flex justify-center">
-      <ThemeProvider>
-      <SampleThemeToggler/>
-      </ThemeProvider>
     </div>
 
   </Tab>
