+++ packages/react/src/hooks/use-theme-toggler.tsx
@@ -1,10 +1,4 @@
-"use client";
-
 import { createContext, ReactNode, useContext, useState } from "react";
-import { Button } from "../components/button";
-import { SunIcon } from "../icons/sun";
-import { MoonIcon } from "../icons/moon";
-import { PlaceholderIcon } from "../icons";
 
 interface ThemeContextType {
   theme: string;
@@ -18,7 +12,7 @@ const ThemeProvider = ({
   themeKey = "theme",
 }: {
   children: ReactNode;
-  themeKey: string;
+  themeKey?: string;
 }) => {
   const [theme, setTheme] = useState<string>("light");
   const switchTheme = (newTheme: string) => {
@@ -43,37 +37,4 @@ const useTheme = () => {
   return context;
 };
 
-interface Themes {
-  theme: string;
-  icon: ReactNode;
-}
-
-interface ThemeToggler {
-  themes: Themes[];
-}
-
-const ThemeToggler = ({
-  themes = [
-    { theme: "light", icon: <SunIcon /> },
-    { theme: "dark", icon: <MoonIcon /> },
-  ],
-}: {
-  themes: Themes[];
-}) => {
-  const { toggleTheme } = useTheme();
-  const [currentIndex, setCurrentIndex] = useState(0);
-
-  const handleClick = () => {
-    const nextIndex = (currentIndex + 1) % themes.length;
-    setCurrentIndex(nextIndex);
-    toggleTheme(themes[nextIndex]?.theme!);
-  };
-
-  return (
-    <Button variant="default-ghost" className="p-2" onClick={handleClick}>
-      {themes[currentIndex]?.icon ?? <PlaceholderIcon />}
-    </Button>
-  );
-};
-
-export { ThemeProvider, useTheme, ThemeToggler };
+export { ThemeProvider, useTheme };
